Drop named slice types, flag's Short name is rune
Named slice types where used for the "automatic" msgpack encoding
and decoding support, as we now call these methods "manually" the
helper types can be dropped.

The Short name of the Flag is now rune instead of string as only
single character names can be used for the short name.